quantmetry_ResNet_api/requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
tensorflow==2.3.1
pillow==8.0.1
python-multipart==0.0.5
pytest==5.2.1